Improve build configuration and documentation for Debug and Release modes
- Add CMake configuration for Debug and Release build types
- Set optimization flags for Release mode (-O3 for GCC/Clang, /O2 for MSVC)
- Update README with detailed build instructions for Debug and Release modes
- Add performance test option in main.cpp
- Improve type casting and error handling in various source files
- Enhance build flexibility and performance optimization

+++ cpp/src/board.cpp
@@ -45,7 +45,7 @@ uint16_t Board::moveLeft(uint16_t row, int& score) {
         }
     }
 
-    return (result[0] << 12) | (result[1] << 8) | (result[2] << 4) | result[3];
+    return static_cast<uint16_t>((result[0] << 12) | (result[1] << 8) | (result[2] << 4) | result[3]);
 }
 
 uint16_t Board::moveRight(uint16_t row, int& score) {
@@ -60,8 +60,8 @@ void Board::initLookupTables() {
     for (int i = 0; i < (1 << 16); ++i) {
         int leftScore = 0;
         int rightScore = 0;
-        leftMoves[i] = moveLeft(i, leftScore);
-        rightMoves[i] = moveRight(i, rightScore);
+        leftMoves[i] = static_cast<uint16_t>(moveLeft(static_cast<uint16_t>(i), leftScore));
+        rightMoves[i] = static_cast<uint16_t>(moveRight(static_cast<uint16_t>(i), rightScore));
         leftScores[i] = leftScore;
         rightScores[i] = rightScore;
     }
